Revert "Generate preview images with node script"
