+++ map-editor/map-editor.js
